uk-pmrv-app-web/src/app/actions/aer/reviewed/additional-info/additional-info.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { ActivatedRoute } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';

import { ActivatedRouteStub, BasePage } from '@testing';
import { KeycloakService } from 'keycloak-angular';

import { CommonActionsStore } from '../../../store/common-actions.store';
import { AerModule } from '../../aer.module';
import { mockStateCompleted } from '../testing/mock-aer-completed';
import { AdditionalInfoComponent } from './additional-info.component';

describe('AdditionalInfoComponent', () => {
  let page: Page;
  let store: CommonActionsStore;
  let component: AdditionalInfoComponent;
  let fixture: ComponentFixture<AdditionalInfoComponent>;

  const route = new ActivatedRouteStub(
    {},
    {},
    {
      groupKey: 'ADDITIONAL_INFORMATION',
    },
  );

  class Page extends BasePage<AdditionalInfoComponent> {
    get heading(): string {
      return this.query<HTMLHeadingElement>('h1').textContent.trim();
    }

    get summaryListValues() {
      return this.queryAll<HTMLDivElement>('.govuk-summary-list__row')
        .map((row) => [row.querySelector('dt'), row.querySelector('dd')])
        .map((pair) => pair.map((element) => element.textContent.trim()));
    }
  }

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      providers: [KeycloakService, { provide: ActivatedRoute, useValue: route }],
      imports: [AerModule, RouterTestingModule],
    }).compileComponents();
  });

  beforeEach(() => {
    store = TestBed.inject(CommonActionsStore);
    store.setState(mockStateCompleted);

    fixture = TestBed.createComponent(AdditionalInfoComponent);
    component = fixture.componentInstance;
    page = new Page(fixture);
    fixture.detectChanges();
    jest.clearAllMocks();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should show summary details', () => {
    expect(page.heading).toEqual('Additional information');
    expect(page.summaryListValues).toHaveLength(11);
    expect(page.summaryListValues).toEqual([
      ['Abbreviation, acronym or terminology', 'Mr'],
      ['Definition', 'Mister'],
      ['Abbreviation, acronym or terminology', 'Ms'],
      ['Definition', 'Miss'],
      ['Additional documents or information', 'No'],
      ['Section', 'Section 1'],
      ['Explanation', 'Explanation 1'],
      ['Section', 'Section 2'],
      ['Explanation', 'Explanation 2'],
      ['Decision status', 'Accepted'],
      ['Notes', 'Notes'],
    ]);
  });
});
